+++ AtraShared/Utils/GameLocationUtils.cs
@@ -215,6 +215,24 @@ public static class GameLocationUtils
         }
     }
 
+    /// <summary>
+    /// Gets all the buildings.
+    /// </summary>
+    /// <returns>IEnumerable of all buildings.</returns>
+    public static IEnumerable<Building> GetBuildings()
+    {
+        foreach (GameLocation? loc in Game1.locations)
+        {
+            if (loc is BuildableGameLocation buildable)
+            {
+                foreach (Building? building in GetBuildings(buildable))
+                {
+                    yield return building;
+                }
+            }
+        }
+    }
+
     private static IEnumerable<GameLocation> YieldInteriorLocations(BuildableGameLocation loc)
     {
         foreach (Building building in loc.buildings)
@@ -228,6 +246,21 @@ public static class GameLocationUtils
                     {
                         yield return nestedLocation;
                     }
+                }
+            }
+        }
+    }
+
+    private static IEnumerable<Building> GetBuildings(BuildableGameLocation loc)
+    {
+        foreach (Building building in loc.buildings)
+        {
+            yield return building;
+            if (building.indoors?.Value is BuildableGameLocation buildable)
+            {
+                foreach (Building interiorBuilding in GetBuildings(buildable))
+                {
+                    yield return interiorBuilding;
                 }
             }
         }