+++ QualityRings/Framework/AbstractCraftingProfession.cs
@@ -4,7 +4,8 @@ namespace QualityRings.Framework;
 internal abstract class AbstractCraftingProfession : Skills.Skill.Profession
 {
 
-    protected AbstractCraftingProfession(CraftingSkill skill, string id) : base(skill, id)
+    protected AbstractCraftingProfession(CraftingSkill skill, string id)
+        : base(skill, id)
     {
     }