+++ QualityRings/Framework/RingQualityManager.cs
@@ -1,4 +1,36 @@
-﻿namespace QualityRings.Framework;
-internal class RingQualityManager
+﻿using System.Runtime.CompilerServices;
+using AtraShared.Utils.Extensions;
+using StardewValley.Objects;
+
+namespace QualityRings.Framework;
+
+/// <summary>
+/// Caches the quality of rings so I'm not bloody parsing them from modData all the time.
+/// </summary>
+internal static class RingQualityManager
 {
+    private const string RingQuality = "atravita.RingQuality";
+
+    private record Holder(int quality);
+
+    private static readonly ConditionalWeakTable<Ring, Holder> QualityMap = new();
+
+    /// <summary>
+    /// Gets the quality for a ring.
+    /// </summary>
+    /// <param name="ring">Ring in question.</param>
+    /// <returns>quality of the ring.</returns>
+    internal static int GetQuality(Ring ring)
+        => QualityMap.TryGetValue(ring, out Holder? val) ? val.quality : ring.modData?.GetInt(RingQuality) ?? 0;
+
+    /// <summary>
+    /// Sets the quality for a ring.
+    /// </summary>
+    /// <param name="ring">Ring in question.</param>
+    /// <param name="quality">quality of the ring.</param>
+    internal static void SetQuality(Ring ring, int quality)
+    {
+        ring.modData?.SetInt(RingQuality, quality);
+        QualityMap.AddOrUpdate(ring, new Holder(quality));
+    }
 }